Convert 'many-buttons' module to directory
