+++ Cargo.toml
@@ -1931,7 +1931,7 @@ wasm = true
 
 [[example]]
 name = "many_buttons"
-path = "examples/stress_tests/many_buttons.rs"
+path = "examples/stress_tests/many_buttons/main.rs"
 doc-scrape-examples = true
 
 [package.metadata.example.many_buttons]
