+++ examples/stress_tests/many_buttons/idle_color.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::{BackgroundColor, Component};
+
+#[derive(Component)]
+pub(crate) struct IdleColor(pub(crate) BackgroundColor);
